+++ packages/esbuild-plugin-react18-example/src/server/constants.ts
@@ -1,0 +1,3 @@
+/** we will replace default colors to test sourceReplacePatterns */
+export const defaultBgColor = "#aaa";
+export const defaultColor = "#555";