+++ packages/esbuild-plugin-react18/CHANGELOG.md
@@ -1,5 +1,13 @@
 # esbuild-plugin-react18
 
+## 0.0.3
+
+### Patch Changes
+
+- e5d3a58: remove data-testid as last item in sourceReplacePatterns. This will reduce the onLoad conflicts significantly as this pattern matches all the js, ts, jsx and tsx files.
+- Fix buildReplacePatterns to work on all esbuild setups
+- e186209: Fix ignorePatterns with contentPatterns
+
 ## 0.0.2
 
 ### Patch Changes